fix(storage): bind recovery namespace transitions
